services/auth/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
anyio==3.6.1
async-timeout==4.0.2
asyncpg==0.26.0
attrs==22.1.0
bcrypt==4.0.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.83.0
fastapi-jwt-auth==0.5.0
frozenlist==1.3.1
h11==0.12.0
httpcore==0.15.0
httptools==0.4.0
httpx==0.23.0
idna==3.3
iniconfig==1.1.1
iso8601==1.0.2
multidict==6.0.2
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
PyJWT==1.7.1
pyparsing==3.0.9
pypika-tortoise==0.1.6
pytest==7.1.3
pytest-aiohttp==1.0.4
pytest-asyncio==0.19.0
python-dotenv==0.21.0
python-jose==3.3.0
pytz==2022.2.1
PyYAML==6.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
tomli==2.0.1
tortoise-orm==0.19.2
typing_extensions==4.3.0
uvicorn==0.18.3
uvloop==0.16.0
watchfiles==0.17.0
websockets==10.3
yarl==1.8.1
